feat(lib/log): log attributes in context (#77)
Adds support for log attributes in the context via `log.WithCtx`. This
reduces a of duplication and allows upstream callers to provide
consistent downstream log attributes.

Included a refactor of xprovider to show the affect.

Also added a "gotest-time" check that verifies all logging keys are in
"snake_case". It uses build tags to avoid performance issues in prod.

task: none