ci(*): notify slack on ci failures (#128)
Added a slack webhook, and running a notify action after every workflow
run on main, and notifying us on those that fail. In the documentation,
t's unclear if `"*"` syntax is supported, and if it doesn't work we'll
need to add which workflows we want to check.

The current webhook uses the #eng channel. If this becomes too noisy, we
can easily move it to #alerts. But resolving failures on `main` should
be high pri, so I kept it in #eng for now.

task: none