fix(e2e): reset anvil loader to block 0 (#343)
The `xprovider` watches a chain's portal from a provided start block. We
set our private anvil chains to start at block 0 (via portal
`DeployInfo`). When we loaded anvil state for the Eigenlayer
deployments, it started from block 12. Then our `xprovider` was watching
for block 0, and never processing any `XBlocks`.

The solution is to just set the anvil pre-loaded chain to start at block
0. Another solution could be to set portal `DeployInfo` once the portal
is actually deployed, but this is a more difficult – as we don't
actually spin up the anvil chains until after setting the `network.json`

task: none