fix(halo/evmengine): allow proper blocks to be built (#364)
Fixes issue where it takes a long time for omni_evm txs to be minded (up
to 35min). The issue is that we need to wait for geth to mine proper
blocks. If we query to early, it returns empty blocks. Configure geth to
build faster `--miner.recommit` and wait for that.

task: none