chore(docs): add docusaurus docs (#384)
This PR introduces docusaurus dev docs.

To preview the docs in your browser: 
- visit https://omni-gamma-beige.vercel.app/ 
- or checkout the branch, cd to docs/site, install deps and run `start`

Four main sections are introduced in these docs, and I'm keen to hear
what people think:

1. Learn - target all audiences with general information about Omni
2. Use - target knowledgeable/developer audience with specifics about
how Omni works (curious for comments from Lazar955 corverroos)
3. Develop - target dApp developers seeking to build with Omni (curious
for comments from kevinhalliday jmozah)
4. Operate - target node runners on how to run a validator (curious for
comments from corverroos jmozah)

I'd vouch to release the site in a minimal version quickly, and then
continue improving and editing through the coming days.

@ttarsi seems as if the PR action for commenting the URL I posted in
this description does not have write permissions for PRs. I tried
editing the settings in the settings for the repo but I appear not to be
permissioned to elevate the rights of github actions agents so that it
can perform that commenting part. So that is the reason that CI step is
failing.

task: https://app.asana.com/0/1206208509925075/1206566996089561/f