chore(e2e/prometheus): update prometheus command (#407)
Until now we were running Prometheus with the default settings (command
flags). With this PR we set those explicitly and enable 2 new flags:
- Instruct Prometheus to scrape and store metrics exemplars
- Disable the Prometheus UI on our staging env and only run it in agent
mode. This way it consumes a lot less resources. The agent mode can be
potentially conditional and on localhost we can run the full Prometheus
UI.

task: none