feat(explorer/api): refactor graphql schema (#980)
This is a WIP refactoring the Explorer GraphQL API starting from the
schema first. To unblock the front-end developers this PR adds a mock
resolver which returns fuzzed data with pagination enabled. Further
refactoring of the backend should not affect the front-end work or if
any changes are required they should be minimal/negligible.

task: https://app.asana.com/0/1206208509925075/1207214156189314/f