chore(halo/voter): attest to multiple versions (#1129)
Add `ConfFinalized` as a confirmation level for all chains. This ensure
we attest to all chains either once (if `ConfFinalized`) or twice if
(`ConfFuzzy`).

Note this only affects halo voter and attest module. Relayer only
streams the expected conf level at this point.

Update `ForkProxy` to support custom `--slots-in-an-epoch` to speed up
finalization from 64 blocks (+1min) to 12sec.

task: none