fix(octane/evmengine): handle unknown payload error (#1238)
Handle edge-case of omni_evm restart after payload was triggered in
which case it may return "Unknown Payload". Currently it gets stuck in
that loop. Now, it will error and propose an empty block.

Also increase genesis validator power to 100 since that is the minimum
stake in the omni staking contract.

task: none