fix(lib): endpoint HeaderByNumber might return nil (#1568)
The result of a call to `eth_getBlockByNumber` can be `nil`, with a
`nil` error. This most often happens when calling the `Pending` block,
but it can happen in other situations (see
[here](https://github.com/ethereum/go-ethereum/blob/ed8fd0ac0919ee954cfa621d16a5d0cfbf30c96d/core/blockchain.go#L929)).
geth [checks for the returned block
](https://github.com/ethereum/go-ethereum/blob/2e8e35f2ada60e8e90aff8f6374ea7cd7da16116/ethclient/ethclient.go#L217-L219)being
`nil` before returning, throwing an error in this case. We should do the
same.

This PR refactors things to use geth's `HeaderByNumber` method, which
will perform the check for us.

issue: none