refactor(docs): cleanup docs structure (#1692)
Goal here was just to make it clearer to see how and where to contribute
to docs. Will help with our current revamp + make future contributions
easier. Addresses:
- previously nested structure that made navigating content files a bit
confusing: `docs/site/docs/..` => `docs/content/..`
- contributing docs which have their own subdir, we'll need to do some
work on these in the future, but we don't need to optimize for this now

issue: none