fix(halo/voter): abort app if persistence fails (#1998)
If voter fails to persist state to disk, then the in-memory state is
inconsistent. This could result in double-signing and possible slashing.

Instead of trying to fix in-memory state via rollbacks or
compare-and-swap type logic, just abort the voter (and shutdown halo).
This indicates a full disk (or loss of validator state file), so best
crash(loop).

This issue was raised by the spearbit audit.

issue: none
(cherry picked from commit 675766f11c533fd29eadf31b2f87616e733ff159)