fix(provider): potential nil pointer dereference (#2120)
This PR addresses a potential panic issue in the node's handling of the
msgs slice in the xblock response. If any message in the msgs slice is
nil, attempting to access msg.ShardID will lead to a panic. In a
Byzantine environment, it’s crucial to consider that a malicious client
could return a compromised block response, which may include nil
messages. Implementing this check helps safeguard the node against such
scenarios, enhancing its resilience against potential attacks.

issue: none