test(contracts/core): added basic smoke tests for preinstalls (#2243)
I added a handful of basic functionality tests purely intended to ensure
each preinstall is functioning and that their bytecode can execute. Most
of the tests involve view function call checks, however some also handle
writes as well. At the time of PR, `BeaconBlockRoots` is the only
preinstall remaining that is untested due to requiring an evm_version,
however, this is not being done as it leads to bytecode changes.

issue: #1399

---------

Co-authored-by: Kevin Halliday <kevin@omni.network>