fix(attest/voter): reset to available if not included (#2424)
Currently, if a validator prosed a vote(s) but they are not included in
the next rounds proposed block, they will only be reset to available if
that (or once subsequent) proposed blocks contain another vote from the
same validator. This creates an inefficiency, where votes are only
rebroadcasted once subsequent votes are included. This could delay
voting by up to `AttestInterval` (1 hour).

It is better to rebroadcast if even no votes from the validator is
included in the next proposal.

issue: none