feat(octane/evmengine): support blob txs (#2498)
Adds minimal support for blob txs.

Only include `BlobCommitments` in consensus chain blocks. This is
required to calculate blob hashes and provide to `engine.NewPayload`.
This adds minimum support for blobs, i.e., that they can be submitted to
EVM and included in blocks without problems.

But we don't include blobs themselves (or proofs) in the consensus
chains. Neither do we provide an API to query them. This limits
consensus chain size by not including blobs. Submitted blobs can
therefore not be queried from the consensus chain. This does render
BlobTxs useless, but this isn't a problem since we don't have a use-case
that require it.

Minimal support for blobs does means that stock standard geth is still
supported and that forking geth to disable blob txs isn't required.

issue: https://github.com/omni-network/ops/issues/613