fix(cli/cmd): update operator init-nodes cmd (#2703)
This PR updates the flags for the operator init-nodes cmd so that the
cmd can be run directly from source without having to install the omni
CLI. This PR adds a new flag **--halo-tag**, which is the halo tag to
use for the new node. This is required because the current
`buildinfo.GitCommit() ` func will return an err if the cmd is not run
from an binary. This PR also binds the current **--feature-flags** to
the init-nodes command. Now we can run the command in CI and locally as
follows

```
cd cli/cmd/omni
go run main.go operator init-nodes --network=staging \
--moniker=foo \
--home=/tmp \
--halo-tag=main \
--feature-flags=evmstaking
```

issue: none