chore(docs): cleanup unused static docs files (#2715)
1. Clean up unused static files in docs. When the preview was showing
our old branding yesterday, it was difficult to know which files were
being used. This removes files that are unused so that it's easier to
know what is live.
2. Moves `static-avs-metadata.json` to be hosted on the docs site, so
that we can deprecate the `static` directory (less is more, want to make
the monorepo less confusing). Still need to make an onchain transaction
so our AVS contract still displays correctly on the eigen frontend
(updates where to pull metadata from) but we can merge this and then
remove the static dir later.

issue: #2712