feat(contracts/core): support user stake delegation (#2681)
Added _initial_ support for user stake delegation to the `Staking`
contract, and updated related Go. I also added support for delegating on
behalf of other users. The omni CLI delegate command was only updated
for the new delegate function interface, so CLI still only supports
self-delegation, for now.

It has yet to be determined what contract changes, if any, are necessary
for unstaking. This contract should not be upgraded yet until the entire
scope is known.

issue: #567