feat(contracts/solve): implemented SolverNetExecutor and AddrUtils (#2794)
SolverNetExecutor was implemented to migrate call execution outside of
the contract solvers approve to manage funds. Because arbitrary calldata
will eventually be executed by the Outbox, it is important that it
cannot be misused to steal solver funds (by calling transfer functions
on tokens solvers have approved). This is done by having the Outbox
leverage solver approvals to send tokens to the Executor, such that the
Executor is outside of those approvals.

address <> bytes32 conversions were also moved into a separate AddrUtils
library.

issue: #2791