feat(cli): add edit validator to cli (#2925)
Adds the following command the cli: `omni operator edit-validator`.

Also adds a e2e test for this.

Switches solidity to signed ints to use -1 as "do not modify". This is
required since the previous approach of setting the existing value
failed, since CommissionRate can only be updated very 24 hours (even to
same value). So we have to identify "nulls".

issue: #2886

---------

Co-authored-by: Zodomo <zodomo@proton.me>