chore(lib/ethclient): close idle connections periodically (#3590)
Add `CloseIdleConnectionsForever` to ethclient. Call it in production
use-cases.

This aims to mitigate the problem of TCP-keep-alive maintaining
long-lived connections to stalled (but alive) RCP servers. This
sometimes occurs with QuickNode, when one of the load-balanced upstream
servers/nodes stalled, but we remain connected to it, requiring a
restart of the affected service to reconnect to a healthy upstream.

By periodically closing idle connections, we get best of both worlds,
TCP-keep-alive improves performance of most queries since TCP connection
doesn't need to be reestablished for each query. But we still reconnect
every 5min ensuring good load balancing over time.

This also required some cleanup:
 - Pass in context in production uses of `ethclient.DialContext`.
 - Remove `ethclient` duplication in `monitor`.

issue: none