fix(halo): relax block validation around upgrades (#3692)
Introduces the `strictMode` flag to the block payload validation which
defines whether the payload should be rejected if it contains no
withdrawals, even though we have some eligible ones. This is required,
because after we upgrade the network to a new binary, the first block it
executes is built by the previous binary, which cannot include the
withdrawals.

This change should be deleted after drake is rolled out.

issue: none