feat(contracts/solve): integrate hyperlane (#3710)
Integrated Hyperlane as an alternative messaging provider to help us
expand to more chains quickly in the short term. The `SolverNetInbox`
can receive fulfillment attestations regardless of if they are routed
via Omni Core or Hyperlane. The `SolverNetOutbox` can be configured to
use either Omni Core or Hyperlane to route fulfillment attestations to
an Inbox on any given chain. There are no breaking changes for the
solver, this is a backwards-compatible additive change. We use a
customized Hyperlane `MailboxClient` contract suited for our use case,
and also use refactored mock contracts to test against to eliminate
dependency issues.

issue: #3741