feat(contracts/solve): remove upgrade gap from MailboxClient (#3945)
SolverNet inbox and outbox upgrades were failing on Omega due to a
storage layout mismatch detected in our upgrade script. This was
introduced by a 48-slot upgrade gap in Hyperlane's contract that we did
not need. I've removed this gap, which realigned the storage layouts,
correcting this problem. We avoided seeing this on staging because we
began Hyperlane testing with a reset of staging, which caused us to
start with the new storage layout.

issue: #3925