feat(solver): enable RPC overrides in solver (#3968)
Solver app itself was not retrieving the RPC overrides, so the Hyperlane
network did not include any Hyperlane chains due to missing endpoints.
I've attempted to bind these RPC overrides to the solver config. This PR
doesn't functionally change the solver as the hyperlane extended network
is not used after creating it, as I want to see if this works first.

issue: none