fix(e2e): post redenom bridge balance (#4361)
Upgrading native bridge on staging was failing because the `l1Deposits`
value was near the max uint256 due to an underflow caused by subtracting
the post-redenominated bridge balance from the OMNI total supply value,
which produced a negative number. Added a check for this edge case so
that the value is correct.

issue: none