feat(e2e): use NOM in place of OMNI (#4389)
Replaced most references of `tokens.OMNI` with `tokens.NOM`, as well as
`addrs.Token` with `addrs.NomToken`. This PR also migrates the native
token definition for Omni Mainnet to NOM, and deprecates use of OMNI in
the solver and fee manager. The only mechanisms left that reference
`tokens.OMNI` are funding mechanisms, and those will be handled in a
later PR.

issue: none